requirements.txt
aniso8601==8.0.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
h5py==2.10.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==1.2.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
opencv-python==4.2.0.34
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pdf2image==1.13.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
scikit-learn==0.22.2.post1
scipy==1.1.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorflow==1.0.1
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
